feat: BSZ-GOV-01-ownership-checks-006 - Split ownership rationale catalog below file-size limit
