+++ internal/ownershipinventory/rationale.go
@@ -124,824 +124,6 @@ func committedOwnerRationales() []OwnerRationaleCard {
 	out = append(out, committedWorkersRationales()...)
 	return out
 }
-func committedAutomationsRationales() []OwnerRationaleCard {
-	return []OwnerRationaleCard{
-		topLevel(
-			"automations",
-			"pkg/services/automations",
-			"Desired/observed trigger instances, cursors/checkpoints, and Work Request generation from clocks or external observations.",
-			"Automation definitions, instance status, and cursor/checkpoint store owned by Automations.",
-			"Reconciles desired specs into long-running trigger instances with start/stop/join/restart policy.",
-			"Factory Runtime (via Work commands), Work Admission, Operator Settings, transports commanding reconcile/list/control.",
-			"Trigger instance and cursor mutations stay inside Automations; Work admission is a cross-owner command.",
-			"Missed triggers, duplicate emissions, cursor recovery, cancellation, and downstream Work rejection are typed Automations facts.",
-		),
-		nested(
-			"automations/cron",
-			"automations",
-			"automations",
-			"pkg/services/automations/internal/services/cron",
-			"Deterministic schedule ticks that command Work admission.",
-			"Schedule/jitter/expiry configuration and tick cursors under Automations store.",
-			"Produces ticks under Reconciliation supervision; no independent daemon outside Automations.",
-			"Automations Reconciliation and Work Admission via Automations root.",
-			"Shares Automations instance/cursor transaction; does not own Work content.",
-			"Invalid schedules and missed ticks surface as Automations status/recovery facts.",
-		),
-		nested(
-			"automations/filesystem_watchers",
-			"automations",
-			"automations",
-			"pkg/services/automations/internal/services/filesystem_watchers",
-			"Watch configured paths, preseed, debounce/coalesce, persist cursor, and command Work.",
-			"Watcher cursors and configured path subscriptions in Automations store.",
-			"Long-running watch instances supervised by Reconciliation.",
-			"Automations root consumers and Work Admission; moves filesystem ingest off Factory Runtime.",
-			"Cursor and subscription mutations are Automations-local; Work creation is a command.",
-			"Watcher restart and cursor recovery remain Automations failure/recovery policy.",
-		),
-		nested(
-			"automations/hosted_sources",
-			"automations",
-			"automations",
-			"pkg/services/automations/internal/services/hosted_sources",
-			"Observe hosted systems, poll/restart/checkpoint, normalize observations, and command Work.",
-			"Hosted-source cursors/checkpoints and secrets resolution handles under Automations.",
-			"Polling/restart lifecycle owned by Automations, not Workers Hosted Runner.",
-			"Automations root and Work Admission; Workers retains only hosted Work execution.",
-			"Observation/cursor state is Automations-local; Work Requests cross the Work boundary.",
-			"Poll/restart and secret-resolution failures classify as Automations source faults.",
-		),
-		nested(
-			"automations/reconciliation",
-			"automations",
-			"automations",
-			"pkg/services/automations/internal/services/reconciliation",
-			"Apply desired automation specs, construct instances, track status, and supervise lifecycle.",
-			"Desired/observed automation instance registry and status in Automations store.",
-			"Owns start/stop/join/restart orchestration for trigger subservices.",
-			"Automations root command surface and trigger subservices.",
-			"Desired-spec application and instance status mutations are one Automations aggregate.",
-			"Restart policy and reconcile conflicts remain Automations recovery behavior.",
-		),
-		nested(
-			"automations/script_pollers",
-			"automations",
-			"automations",
-			"pkg/services/automations/internal/services/script_pollers",
-			"Supervise polling commands/sources that emit canonical Work Requests.",
-			"Poller cursors/checkpoints under Automations store.",
-			"Timeout/restart supervision under Reconciliation.",
-			"Automations root and Work Admission.",
-			"Cursor and poller status are Automations-local; emitted Work crosses owners by command.",
-			"Timeout and parse failures remain Automations poller recovery facts.",
-		),
-	}
-}
-func committedFactoryDefinitionsRationales() []OwnerRationaleCard {
-	return []OwnerRationaleCard{
-		topLevel(
-			"factory_definitions",
-			"pkg/services/factory_definitions",
-			"Authored Factory aggregate catalog: validation, persistence, versioning, activation, and packaged installation.",
-			"Named Factory roots, authored layouts, .current-factory pointer, snapshots, and packaged assets.",
-			"No autonomous background daemon; activation is request-driven through Factory Sessions.",
-			"Factory Sessions, System Bootstrap, Automations (trigger specs), transports, Wire construction.",
-			"Catalog/authoring/compilation/validation/snapshot/distribution share one Definition aggregate transaction.",
-			"Malformed layouts, activation conflicts, and persistence failures remain Definition-owned typed facts.",
-		),
-		nested(
-			"factory_definitions/authoring_layout",
-			"factory_definitions",
-			"factory_definitions",
-			"pkg/services/factory_definitions/internal/services/authoring_layout",
-			"Parse/render split authored layouts and atomically create/replace/flatten/expand one Factory aggregate.",
-			"Authored layout files under Definition-owned filesystem datastore.",
-			"Request-driven authoring operations; no independent daemon.",
-			"Factory Definitions root and Definition-owned transports.",
-			"Layout writes participate in the Definition aggregate transaction.",
-			"Parse/render and layout conflicts are Definition authoring failures.",
-		),
-		nested(
-			"factory_definitions/catalog",
-			"factory_definitions",
-			"factory_definitions",
-			"pkg/services/factory_definitions/internal/services/catalog",
-			"List/delete/resolve named/project/global Factories and atomically manage the current pointer.",
-			"Named Factory catalog and .current-factory pointer.",
-			"Catalog mutations are command-driven; no background reconciler.",
-			"Factory Definitions root, System Bootstrap, Factory Sessions activation.",
-			"Catalog and current-pointer updates are atomic within Definitions.",
-			"Missing/conflicting names and pointer races are Definition catalog failures.",
-		),
-		nested(
-			"factory_definitions/compilation",
-			"factory_definitions",
-			"factory_definitions",
-			"pkg/services/factory_definitions/internal/services/compilation",
-			"Convert authored/canonical source into one normalized LoadedFactorySource without executing it.",
-			"Compiled effective definition artifacts derived from authored store.",
-			"Compile on command; Runtime executes, Definitions does not.",
-			"Factory Definitions root and Runtime semantic-validation callers via root.",
-			"Compilation outputs are Definition-owned; Runtime does not mutate authored store.",
-			"Compile/reference-resolution failures stay on the Definition boundary.",
-		),
-		nested(
-			"factory_definitions/distribution",
-			"factory_definitions",
-			"factory_definitions",
-			"pkg/services/factory_definitions/internal/services/distribution",
-			"Built-in package catalog, packaged installation, and scaffold creation producing Factory aggregates.",
-			"Packaged Factory assets and installation outputs in Definition store.",
-			"Install/scaffold commands; no marketplace lifecycle in this program.",
-			"System Bootstrap, CLI/HTTP Definition adapters, Factory Definitions root.",
-			"Installation writes the same Definition aggregate transactionally.",
-			"Install/scaffold failures roll back or report Definition distribution faults.",
-		),
-		nested(
-			"factory_definitions/invocation_policy",
-			"factory_definitions",
-			"factory_definitions",
-			"pkg/services/factory_definitions/internal/services/invocation_policy",
-			"Definition-owned interpretation of authored Factory behavior at invocation time: decision envelopes, interpolation, output shaping, work-type selection, quorum lineage, work propagation, workstation execution limits, and TTS observability identity.",
-			"Stateless invocation-time policy derivations over authored worker, workstation, and packaged configuration.",
-			"Policy interpretation runs on command during worker and workstation invocation; no independent daemon.",
-			"Factory Definitions root, Factory Sessions invocation, and Workers dispatch consumers via Definitions root.",
-			"Invocation policy does not mutate peer stores; interpretation outputs are Definition-owned values consumed at invocation.",
-			"Malformed authored policy and unsupported invocation contracts surface as Definition-owned rejection facts.",
-		),
-		nested(
-			"factory_definitions/snapshots_portability",
-			"factory_definitions",
-			"factory_definitions",
-			"pkg/services/factory_definitions/internal/services/snapshots_portability",
-			"Capture detached snapshots and prepare/materialize bundled assets for export/import/replay use.",
-			"Editable/version snapshots and bundled portable assets.",
-			"Snapshot/export/import are request-driven operations.",
-			"Recordings replay consumers via Definition compile/import root operations; Definition transports.",
-			"Snapshot materialization stays inside Definitions; Recordings owns replay plans.",
-			"Portability validation and materialization failures are Definition-owned.",
-		),
-		nested(
-			"factory_definitions/validation",
-			"factory_definitions",
-			"factory_definitions",
-			"pkg/services/factory_definitions/internal/services/validation",
-			"Structural, topology, pre-persist, tool, and workstation compatibility validation policy.",
-			"Validation policy and diagnostics over Definition aggregate; may query Runtime root for orchestration semantics.",
-			"Validation runs on command before persist/activation.",
-			"Factory Definitions root and authoring/activation callers.",
-			"Validation does not mutate peer stores; persist remains Definition transaction.",
-			"Validation diagnostics are Definition rejection facts, not Runtime failures.",
-		),
-	}
-}
-func committedFactoryRuntimeRationales() []OwnerRationaleCard {
-	return []OwnerRationaleCard{
-		topLevel(
-			"factory_runtime",
-			"pkg/services/factory_runtime",
-			"Per-session orchestration, execution state, instance mechanics, dispatch planning, and checkpoints.",
-			"Runtime checkpoints, submission/result/dispatch buffers, and live instance state.",
-			"Owns run handle pause/resume/stop/replacement/recovery for one leased execution instance.",
-			"Factory Sessions (lease/start), Workers (dispatch), Work, Recordings (event candidates), Visualization (sanitized observations).",
-			"Runtime instance and checkpoint mutations are Runtime-local; peer Work/Workers/Recordings interactions are commands/events.",
-			"Crash recovery uses Runtime checkpoints plus Recordings history; orchestration failures normalize at Runtime root.",
-		),
-		nested(
-			"factory_runtime/checkpoint_recovery",
-			"factory_runtime",
-			"factory_runtime",
-			"pkg/services/factory_runtime/internal/services/checkpoint_recovery",
-			"Capture/load versioned execution checkpoints and restore orchestration state idempotently.",
-			"Runtime-owned checkpoint store coordinated with Recordings history ownership.",
-			"Recovery commands after durability decision; not an independent deployable yet.",
-			"Factory Runtime root and Instance Host.",
-			"Checkpoint write/load is Runtime-local; canonical history remains Recordings.",
-			"Incompatible checkpoints and restore conflicts are Runtime recovery failures.",
-		),
-		nested(
-			"factory_runtime/dispatch_planning",
-			"factory_runtime",
-			"factory_runtime",
-			"pkg/services/factory_runtime/internal/services/dispatch_planning",
-			"Convert enabled transitions into stable dispatch intents and accept correlated results.",
-			"In-flight dispatch intent/outbox buffers owned by Runtime.",
-			"Publishes intents after Workers command/result contract lock.",
-			"Workers execution and Factory Runtime Instance Host.",
-			"Intent lifecycle is Runtime-local; Workers owns execution attempts.",
-			"Duplicate/correlated-result conflicts recover inside Runtime outbox policy.",
-		),
-		nested(
-			"factory_runtime/instance_host",
-			"factory_runtime",
-			"factory_runtime",
-			"pkg/services/factory_runtime/internal/services/instance_host",
-			"Construct inert instances from lease/start data and manage run handle lifecycle.",
-			"Live instance handles and runtime-owned artifact closure metadata.",
-			"Start/stop/pause/resume/replace/recover for one Runtime instance.",
-			"Factory Sessions placement/lease commands and Orchestration.",
-			"Instance handle mutations stay in Runtime; Session desired lifecycle stays in Sessions.",
-			"Replacement and recovery failures remain Runtime instance faults.",
-		),
-		nested(
-			"factory_runtime/orchestration",
-			"factory_runtime",
-			"factory_runtime",
-			"pkg/services/factory_runtime/internal/services/orchestration",
-			"Select orchestration kind, compile/validate/preview/execute/resume strategy semantics, encode checkpoints.",
-			"Compiled orchestration strategy state for the active instance.",
-			"Execute/resume under Instance Host; Petri/JavaScript are internal variants.",
-			"Factory Runtime root, Definition semantic validation callers via Runtime root.",
-			"Strategy execution shares Runtime instance transaction; authored Factory config remains Definitions.",
-			"Strategy validation and execution faults normalize through Runtime orchestration.",
-		),
-	}
-}
-func committedFactorySessionsRationales() []OwnerRationaleCard {
-	return []OwnerRationaleCard{
-		topLevel(
-			"factory_sessions",
-			"pkg/services/factory_sessions",
-			"Session identity, desired lifecycle, placement, recovery, discovery, invocation coordination, and session response streams.",
-			"Live session registry, session control store, and session-owned persistence coordination (canonical history delegated to Recordings).",
-			"Owns desired lifecycle reconcile, live selection/activation, and durable start/resume/control.",
-			"Transports, Factory Runtime, Work, Workers, Provider Sessions, Models, Visualization, Recordings.",
-			"Session aggregate mutations are Sessions-local; peer services are commanded through roots.",
-			"Placement/activation conflicts, durable control failures, and timeout/cancellation are Sessions recovery facts.",
-		),
-		nested(
-			"factory_sessions/durable_execution",
-			"factory_sessions",
-			"factory_sessions",
-			"pkg/services/factory_sessions/internal/services/durable_execution",
-			"Durable start/resume/control/inspection and session-owned persistence coordination.",
-			"Session durable execution identity/persistence handles; delegates canonical history/artifacts to Recordings.",
-			"Durable control lifecycle for a session.",
-			"Factory Sessions root, Live Runtime, Recordings root.",
-			"Session persistence coordination is Sessions-local; ledger append is Recordings command.",
-			"Durable control and resume conflicts recover in Durable Execution.",
-		),
-		nested(
-			"factory_sessions/identity",
-			"factory_sessions",
-			"factory_sessions",
-			"pkg/services/factory_sessions/internal/services/identity",
-			"Normalize/discover targets, derive logical keys, resolve folders, registry-aware identity lookup.",
-			"Logical session keys and folder/identity resolution metadata.",
-			"Identity resolution is request-driven under Sessions.",
-			"Factory Sessions root and nested Live Runtime/Durable Execution.",
-			"Identity derivation shares Sessions aggregate authority.",
-			"Ambiguous/missing identity targets are Sessions identity failures.",
-		),
-		nested(
-			"factory_sessions/invocation",
-			"factory_sessions",
-			"factory_sessions",
-			"pkg/services/factory_sessions/internal/services/invocation",
-			"Resolve invocation input, command Work, observe completion, enforce timeout/cancellation, record telemetry.",
-			"Per-session invocation observation state.",
-			"Timeout/cancellation lifecycle for one session invocation.",
-			"Work root, Factory Sessions root, Recordings telemetry consumers.",
-			"Invocation coordination is Sessions-local; Work admission/state are Work commands.",
-			"Timeouts and cancellations are Sessions invocation recovery facts.",
-		),
-		nested(
-			"factory_sessions/live_runtime",
-			"factory_sessions",
-			"factory_sessions",
-			"pkg/services/factory_sessions/internal/services/live_runtime",
-			"Open/list/get/snapshot/pause/resume/close; sole owner of live registry, selection, activation lock, cleanup.",
-			"Live session registry and mutable live-runtime aggregate.",
-			"Live activation lock and cleanup lifecycle.",
-			"Factory Sessions root, Runtime Opening, Durable Execution, Response Stream.",
-			"Live registry mutations are exclusive to Live Runtime inside Sessions.",
-			"Activation races and cleanup failures recover under Live Runtime.",
-		),
-		nested(
-			"factory_sessions/response_stream",
-			"factory_sessions",
-			"factory_sessions",
-			"pkg/services/factory_sessions/internal/services/response_stream",
-			"Allocate session event stores, publish/complete, filter subscriptions, enforce retention/backpressure, track cursors.",
-			"Ephemeral session response-event stores and cursors.",
-			"Publish/complete and subscription lifecycle for a session.",
-			"Factory Sessions root, transports/Visualization presentation consumers.",
-			"Response-store mutations are Sessions-local; canonical Factory events belong to Recordings.",
-			"Backpressure/drop and retention faults are Response Stream recovery policy.",
-		),
-		nested(
-			"factory_sessions/runtime_opening",
-			"factory_sessions",
-			"factory_sessions",
-			"pkg/services/factory_sessions/internal/services/runtime_opening",
-			"Coordinate peer root services and return a lifecycle plan without importing peer implementations.",
-			"No independent datastore; produces lifecycle plan over already-constructed roles.",
-			"Opening/binding lifecycle after root binding contract stabilizes.",
-			"Factory Sessions root and peer service roots (Definitions, Runtime, Workers, Models, Recordings).",
-			"Does not own peer transactions; only plans Sessions-owned activation steps.",
-			"Binding/plan failures surface as Sessions opening faults without peer downcasts.",
-		),
-	}
-}
-func committedFactoryVisualizationRationales() []OwnerRationaleCard {
-	return []OwnerRationaleCard{
-		topLevel(
-			"factory_visualization",
-			"pkg/services/factory_visualization",
-			"Request-activated visualization lifecycle, live projection, response/event presentation, queues, backpressure, final-write coordination.",
-			"Visualization cursors, output queues, and presentation state for an activated request.",
-			"Inert by default; Start/Stop/Wait/cleanup only when request parameters activate it.",
-			"Transports (parameters/codecs/sinks), Recordings query/subscription, sanitized Runtime observations.",
-			"Activation and presentation mutations are Visualization-local; Recordings/Runtime remain peer roots.",
-			"Backpressure drops, activation failures, and drain errors are Visualization recovery facts.",
-		),
-		nested(
-			"factory_visualization/activation_lifecycle",
-			"factory_visualization",
-			"factory_visualization",
-			"pkg/services/factory_visualization/internal/services/activation_lifecycle",
-			"Interpret approved request options, bind session/runtime, own Start/Stop/Wait/cleanup.",
-			"Activation binding state for the request.",
-			"Request-driven visualization lifecycle.",
-			"Factory Visualization root and transport adapters.",
-			"Activation state is Visualization-local.",
-			"Failed binds and unclean shutdowns are activation recovery faults.",
-		),
-		nested(
-			"factory_visualization/live_view_projection",
-			"factory_visualization",
-			"factory_visualization",
-			"pkg/services/factory_visualization/internal/services/live_view_projection",
-			"Retain cursor/events, subscribe retained-then-live, consume Runtime observations and Recordings queries, emit View.",
-			"Live view cursor/event retention owned by Visualization.",
-			"Subscription lifecycle under activation.",
-			"Recordings projection/query root and Runtime observation contract.",
-			"View projection state is Visualization-local; canonical projections stay in Recordings.",
-			"Reconnect/gap handling follows Visualization cursor policy.",
-		),
-		nested(
-			"factory_visualization/response_event_presentation",
-			"factory_visualization",
-			"factory_visualization",
-			"pkg/services/factory_visualization/internal/services/response_event_presentation",
-			"Best-effort/lossless queues, bounded drop/backpressure, exclusive final writes, drain.",
-			"Output queues and presentation buffers.",
-			"Queue drain lifecycle tied to activation.",
-			"Transport sinks/codecs and Visualization root.",
-			"Queue mutations are Visualization-local.",
-			"Drop/backpressure and final-write conflicts are presentation recovery facts.",
-		),
-	}
-}
-func committedModelsRationales() []OwnerRationaleCard {
-	return []OwnerRationaleCard{
-		topLevel(
-			"models",
-			"pkg/services/models",
-			"Opaque runtime scopes, model catalog, assets, supervised hosting/readiness, nested leases, and inference.",
-			"Model catalog/readiness store, asset cache layout, supervised runtime slots and leases.",
-			"Process-scoped service is inert until commanded; hosts manage unload/eviction/shutdown.",
-			"Workers, Factory Sessions, Operator Settings, Models HTTP/CLI adapters.",
-			"Catalog/assets/host/lease/inference share Models process scope; peers never see host internals.",
-			"Pull/host/lease/inference failures normalize at Models root with typed readiness/availability facts.",
-		),
-		nested(
-			"models/assets",
-			"models",
-			"models",
-			"pkg/services/models/internal/services/assets",
-			"Resolve source/cache layout, pull, verify, classify, remove, and report immutable asset readiness.",
-			"Asset cache and verification metadata.",
-			"Pull/remove operations; no independent public lifecycle.",
-			"Models Catalog/Inference via Models root.",
-			"Asset mutations are Models-local.",
-			"Pull/verify failures classify as Models asset readiness faults.",
-		),
-		nested(
-			"models/catalog",
-			"models",
-			"models",
-			"pkg/services/models/internal/services/catalog",
-			"Canonical model identity, configured operation/binding discovery, source metadata, list/get, readiness projection.",
-			"Model catalog and configured binding metadata.",
-			"Catalog queries/commands; readiness projection without hosting.",
-			"Models root consumers including Workers and direct `you models` flows.",
-			"Catalog identity is Models-local; Providers owns external provider IDs.",
-			"Unknown model/operation lookups are Catalog rejection facts.",
-		),
-		nested(
-			"models/inference",
-			"models",
-			"models",
-			"pkg/services/models/internal/services/inference",
-			"Resolve scoped model operation, acquire lease, load/reuse handle, invoke, release, normalize artifacts/failures.",
-			"Invocation-time handle/lease usage over Runtime Host slots.",
-			"Acquire/release around one inference operation.",
-			"Workers and Models transports via Models root.",
-			"Inference does not own catalog/host stores; it uses Host leases transactionally for one call.",
-			"Invocation and artifact-export failures normalize through Inference.",
-		),
-		nested(
-			"models/runtime_host",
-			"models",
-			"models",
-			"pkg/services/models/internal/services/runtime_host",
-			"Own supervised runtime slots, health/readiness, handle reuse, idle unload, resource-pressure eviction, shutdown.",
-			"Supervised process/runtime slot registry.",
-			"Host supervision lifecycle including idle unload and shutdown.",
-			"Models Inference and Leases nested subservice.",
-			"Slot mutations are exclusive to Runtime Host (+ nested Leases).",
-			"Eviction and supervision faults recover inside Runtime Host.",
-		),
-		nested(
-			"models/runtime_host/leases",
-			"models",
-			"models/runtime_host",
-			"pkg/services/models/internal/services/runtime_host/internal/services/leases",
-			"Capacity reservation, holder identity, acquisition/release/expiry, contention over Host-owned slots.",
-			"Lease records over Runtime Host slots.",
-			"Lease acquire/release/expiry lifecycle nested under Host.",
-			"Models Inference via Runtime Host.",
-			"Leases share Host atomic capacity invariant; not a sibling top-level cut.",
-			"Contention/expiry failures are lease recovery facts under Host.",
-		),
-		nested(
-			"models/runtime_scopes",
-			"models",
-			"models",
-			"pkg/services/models/internal/services/runtime_scopes",
-			"Open/close opaque scopes for detached Factory-session model configuration without constructing another Service.",
-			"Scope binding registry rejecting stale/foreign references.",
-			"Open/close scope lifecycle; does not construct hosts/pullers.",
-			"Factory Sessions and Models root callers.",
-			"Scope registry is Models-local and process-scoped.",
-			"Stale/foreign scope references are typed scope faults.",
-		),
-	}
-}
-func committedOperatorSettingsRationales() []OwnerRationaleCard {
-	return []OwnerRationaleCard{
-		topLevel(
-			"operator_settings",
-			"pkg/services/operator_settings",
-			"Operator-owned configuration, backend identity, defaults, provider/worker presets, precedence, semantic updates, atomic persistence.",
-			"Authoritative operator settings document store.",
-			"Load/update/persist on command; no autonomous reconciler.",
-			"Factory Sessions, Workers, Models, System Bootstrap, CLI, Wire.",
-			"Document and resolution share one atomic operator document aggregate.",
-			"Malformed/unsupported/conflict/persistence failures are Operator Settings typed facts.",
-		),
-		nested(
-			"operator_settings/document",
-			"operator_settings",
-			"operator_settings",
-			"pkg/services/operator_settings/internal/services/document",
-			"Load, decode, normalize, validate, identity/version, inventory, semantic update, and atomically persist the operator document.",
-			"Operator settings document files/store.",
-			"Atomic persist lifecycle for document updates.",
-			"Operator Settings Resolution and root consumers.",
-			"Document writes are the Operator Settings transaction.",
-			"Decode/validation/persist failures stay on Document.",
-		),
-		nested(
-			"operator_settings/resolution",
-			"operator_settings",
-			"operator_settings",
-			"pkg/services/operator_settings/internal/services/resolution",
-			"Apply backend scope, environment, defaults, presets, and invocation overrides into an immutable effective selection.",
-			"No independent store; reads Document and environment/effect ports.",
-			"Pure resolution on command.",
-			"Workers/Models/Sessions via Operator Settings root; may query Providers for canonical IDs.",
-			"Resolution does not mutate Document; callers persist via Document commands.",
-			"Conflict/unsupported override failures are Resolution facts.",
-		),
-	}
-}
-func committedProviderSessionsRationales() []OwnerRationaleCard {
-	return []OwnerRationaleCard{
-		topLevel(
-			"provider_sessions",
-			"pkg/services/provider_sessions",
-			"Secure resolution and read-only inspection of provider-owned persisted session artifacts, transcripts, reasoning, tools, parse errors, and usage.",
-			"Resolved provider-session and transcript projections; no provider execution store.",
-			"Read-only resolve/inspect lifecycle; no execution daemon.",
-			"Factory Sessions, Workers (carry SessionRef), HTTP Provider Sessions adapters.",
-			"Readers share Provider Sessions root transaction for one SessionRef resolve; Providers owns SessionRef minting.",
-			"Secure path/parse failures and unknown SessionRef errors are Provider Sessions facts.",
-		),
-		nested(
-			"provider_sessions/codex_reader",
-			"provider_sessions",
-			"provider_sessions",
-			"pkg/services/provider_sessions/internal/services/codex_reader",
-			"Secure JSONL rollout discovery, timestamped layouts, symlink containment, native parsing, normalized detail projection.",
-			"Codex transcript discovery/projection state for a resolve.",
-			"Read-only resolve path for Codex SessionRefs.",
-			"Provider Sessions root registry.",
-			"Reader-local parse projection; no cross-owner writes.",
-			"Symlink/path and parse errors normalize through Codex Reader.",
-		),
-		nested(
-			"provider_sessions/cursor_reader",
-			"provider_sessions",
-			"provider_sessions",
-			"pkg/services/provider_sessions/internal/services/cursor_reader",
-			"Cursor directory/store resolution, SQLite access, protobuf/blob decoding, transcript reconstruction, usage projection.",
-			"Cursor store resolution/projection state for a resolve.",
-			"Read-only resolve path for Cursor SessionRefs.",
-			"Provider Sessions root registry.",
-			"Reader-local decode projection; no provider execution.",
-			"Store/decode failures normalize through Cursor Reader.",
-		),
-	}
-}
-func committedProvidersRationales() []OwnerRationaleCard {
-	return []OwnerRationaleCard{
-		topLevel(
-			"providers",
-			"pkg/services/providers",
-			"Provider catalog/enumeration, identity, availability, capabilities, one normalized execution attempt, native adapters, SessionRef extraction.",
-			"Provider catalog descriptors and availability/capability snapshots; execution is attempt-scoped.",
-			"Catalog refresh and one-shot Execute attempts; no Work retry policy.",
-			"Workers (selection/retry), Operator Settings (validation), transports (`you providers list`), Provider Sessions (SessionRef).",
-			"Catalog identity and one Execute attempt are Providers-local; Workers owns dispatch retry/scheduling.",
-			"Availability/prerequisite and native execution failures classify at Providers; Workers decides retry.",
-		),
-		nested(
-			"providers/catalog",
-			"providers",
-			"providers",
-			"pkg/services/providers/internal/services/catalog",
-			"Canonical provider IDs/aliases, deterministic list/get, metadata, availability/prerequisite probes, capability snapshots.",
-			"Provider catalog and probe snapshots.",
-			"List/get/refresh on command.",
-			"Providers root, Operator Settings validation, Workers selection data.",
-			"Catalog mutations/snapshots are Providers-local.",
-			"Probe and unknown-ID failures are Catalog facts.",
-		),
-		nested(
-			"providers/execution",
-			"providers",
-			"providers",
-			"pkg/services/providers/internal/services/execution",
-			"Select native adapter and perform one normalized provider attempt with stream decode, progress, diagnostics, optional SessionRef.",
-			"Attempt-scoped execution state; no durable Work store.",
-			"One Execute attempt lifecycle.",
-			"Workers via Providers root; Provider Sessions resolves emitted SessionRef.",
-			"Attempt boundary is Providers-local; Workers owns multi-attempt policy.",
-			"Native adapter and failure-normalization faults are Execution facts.",
-		),
-	}
-}
-func committedRecordingsRationales() []OwnerRationaleCard {
-	return []OwnerRationaleCard{
-		topLevel(
-			"recordings",
-			"pkg/services/recordings",
-			"Canonical event acceptance, ordering, replay, recording lifecycle, artifacts/export, subscriptions, and query projections.",
-			"Canonical session event store, artifact store, and projection store.",
-			"Recorder capture/flush lifecycle plus replay/delivery; projections reduce canonical facts.",
-			"Factory Sessions, Factory Runtime, Workers, Visualization, HTTP/CLI dashboard/history consumers.",
-			"Ledger append/projection rebuild share Recordings canonical identity; peers emit facts, they do not write the ledger.",
-			"Gap/cursor/retention, replay divergence, and recorder crash/finalization are Recordings recovery concerns.",
-		),
-		nested(
-			"recordings/artifacts_export",
-			"recordings",
-			"recordings",
-			"pkg/services/recordings/internal/services/artifacts_export",
-			"Persist replay artifacts and build/validate/decode/atomically export privacy-bounded portable recordings.",
-			"Artifact store and portable export packages.",
-			"Export/materialize on command; atomic write policy.",
-			"Recordings Replay and public export consumers.",
-			"Artifact writes are Recordings-local and privacy-bounded.",
-			"Export validation and atomic-write failures are Artifacts/Export faults.",
-		),
-		nested(
-			"recordings/canonical_ledger",
-			"recordings",
-			"recordings",
-			"pkg/services/recordings/internal/services/canonical_ledger",
-			"Validate, sequence, idempotently append, retain, and subscribe to canonical Factory events.",
-			"Canonical session event store with cursor/gap/retention policy.",
-			"Append/subscribe lifecycle for canonical facts.",
-			"Runtime/Workers/Sessions emitters and Projection/Replay consumers via Recordings root.",
-			"Ledger append is the Recordings source-of-truth transaction.",
-			"Idempotency, gap, and retention conflicts recover in Canonical Ledger.",
-		),
-		nested(
-			"recordings/projection_query",
-			"recordings",
-			"recordings",
-			"pkg/services/recordings/internal/services/projection_query",
-			"Deterministically reduce canonical events into world-state and customer/dashboard views; validate reconnect projections.",
-			"Projection store derived from canonical ledger.",
-			"Projection rebuild/query lifecycle private to Recordings.",
-			"Visualization, Sessions, dashboard transports via Recordings root.",
-			"Projection rebuild is Recordings-local; not a separate deployable in this program.",
-			"Reconnect/projection divergence faults are Projection/Query recovery facts.",
-		),
-		nested(
-			"recordings/recording_lifecycle",
-			"recordings",
-			"recordings",
-			"pkg/services/recordings/internal/services/recording_lifecycle",
-			"Select session recording target, capture events, periodic/final flush, error accumulation, crash/finalization.",
-			"Recorder target and flush state for a session recording.",
-			"Capture/flush/finalization lifecycle.",
-			"Factory Sessions/Runtime emitters via Recordings root.",
-			"Recorder state is Recordings-local around canonical append.",
-			"Crash/final flush failures recover in Recording Lifecycle.",
-		),
-		nested(
-			"recordings/replay",
-			"recordings",
-			"recordings",
-			"pkg/services/recordings/internal/services/replay",
-			"Load/validate/hydrate historical recordings, deterministic clock, replay plans/hooks, delivery timing, divergence detection.",
-			"Historical recording loads and replay plan handles.",
-			"Replay delivery lifecycle with deterministic timing.",
-			"Dashboard/CLI replay consumers and Artifacts/Export.",
-			"Replay plans are Recordings-owned; peers do not import replay implementations.",
-			"Divergence and hydrate failures are Replay recovery facts.",
-		),
-	}
-}
-func committedSystemInitializationRationales() []OwnerRationaleCard {
-	return []OwnerRationaleCard{
-		topLevel(
-			"system_initialization",
-			"pkg/services/system_initialization",
-			"Customer-invoked initialization workflow coordinating Operator Settings and Factory Definition without owning their stores.",
-			"No independent datastore; observes initialization outcome only.",
-			"Single Initialize command with idempotent/partial-failure reporting.",
-			"CLI/HTTP `you init` adapters; commands Operator Settings and Factory Definitions roots.",
-			"Cross-owner ordering only; Settings and Definitions retain their own transactions/rollback.",
-			"Partial failure/rollback reporting is Bootstrap-owned without absorbing peer stores.",
-		),
-	}
-}
-func committedWorkRationales() []OwnerRationaleCard {
-	return []OwnerRationaleCard{
-		topLevel(
-			"work",
-			"pkg/services/work",
-			"Work Request admission, secure content staging/materialization, invocation/return policy, lineage, state commands, and detached queries.",
-			"Work and dispatch store plus staged content references.",
-			"Admission and state commands; no scheduling daemon (Automations owns triggers).",
-			"Automations, Factory Runtime, Factory Sessions, Workers, Recordings, Definitions, transports.",
-			"Admission/content/state share Work identity and mutation invariants.",
-			"Acceptance/rejection, staging/materialization, and state-command failures are Work facts.",
-		),
-		nested(
-			"work/admission",
-			"work",
-			"work",
-			"pkg/services/work/internal/services/admission",
-			"Normalize, validate, deduplicate, authorize, and accept Work Requests; emit acceptance/rejection facts.",
-			"Accepted Work Request records in Work store.",
-			"Admission is command-driven.",
-			"Automations and public protocols via Work root; Runtime consumes admitted Work.",
-			"Acceptance is the Work intake transaction.",
-			"Validation/authorization rejections are Admission facts.",
-		),
-		nested(
-			"work/content_materialization",
-			"work",
-			"work",
-			"pkg/services/work/internal/services/content_materialization",
-			"Resolve staged/local/remote/data-URL content with cache, SSRF, size, cleanup, and cancellation policy.",
-			"Materialization cache and resolved content handles.",
-			"Resolve/cleanup lifecycle for content fetches.",
-			"Workers and Work root consumers needing immutable content.",
-			"Materialization policy is Work-local; Workers do not import Work modules.",
-			"SSRF/size/cancellation failures are Materialization facts.",
-		),
-		nested(
-			"work/content_staging",
-			"work",
-			"work",
-			"pkg/services/work/internal/services/content_staging",
-			"Securely persist submitted content, issue/verify opaque references, enforce expiry and owned-path policy.",
-			"Staged content reference store and owned paths.",
-			"Stage/expire/cleanup lifecycle.",
-			"Work Admission and Materialization via Work root.",
-			"Staging writes share Work content transaction.",
-			"Expiry and path-policy violations are Staging facts.",
-		),
-		nested(
-			"work/state_access",
-			"work",
-			"work",
-			"pkg/services/work/internal/services/state_access",
-			"Adapt Factory Sessions root for submit/move/read snapshot operations and detached list/get/move results.",
-			"Detached Work query/projection results over Work store; never exposes Runtime/Petri state.",
-			"Read/move commands without peer store access.",
-			"Transports and peer services via Work root; may query Sessions/Recordings roots.",
-			"State Access does not write Sessions/Recordings stores.",
-			"Not-found and move conflicts are State Access facts.",
-		),
-	}
-}
-func committedWorkersRationales() []OwnerRationaleCard {
-	return []OwnerRationaleCard{
-		topLevel(
-			"workers",
-			"pkg/services/workers",
-			"Runner selection/assembly, workstation behavior, capacity, dispatch execution policy, retry, and Work results.",
-			"Worker registry/execution store and immutable runtime assembly artifacts.",
-			"Dispatch execution and runner lifecycles; Providers owns native provider attempts.",
-			"Factory Runtime dispatch, Factory Sessions, Models, Providers, Work, Recordings.",
-			"Execution/retry policy is Workers-local; Providers Execute is one attempt command.",
-			"Dispatch/retry/cancellation failures normalize as Workers execution facts.",
-		),
-		nested(
-			"workers/runners",
-			"workers",
-			"workers",
-			"pkg/services/workers/internal/services/runners",
-			"Common private Runner contract/registry implemented by agent/script/inference/hosted runners.",
-			"Runner registry identities/capabilities; execution state lives in runner implementations.",
-			"Registry selection for one dispatch; runners own attempt execution.",
-			"Runtime Assembly and Workstation Execution.",
-			"Registry is Workers-private; provider adapters live under Providers.",
-			"Unknown runner/capability mismatches are registry rejection facts.",
-		),
-		nested(
-			"workers/runners/agent",
-			"workers",
-			"workers/runners",
-			"pkg/services/workers/internal/services/runners/internal/services/agent",
-			"Implement Runner for agent harness/tools, retries, stop tokens, decision envelopes, terminal results.",
-			"Agent attempt execution state under Workers.",
-			"Agent runner attempt lifecycle.",
-			"Workers Runner registry; may command Providers/Models through Workers policy.",
-			"Agent attempt is Workers-local; Providers owns native provider attempt details.",
-			"Harness/retry/stop-token failures are Agent Runner facts.",
-		),
-		nested(
-			"workers/runners/hosted",
-			"workers",
-			"workers/runners",
-			"pkg/services/workers/internal/services/runners/internal/services/hosted",
-			"Implement Runner for hosted execution request/result, remote lifecycle observation, cancellation, normalized outcome.",
-			"Hosted execution attempt state under Workers.",
-			"Hosted runner attempt lifecycle (not hosted polling).",
-			"Workers Runner registry; Automations Hosted Sources remain separate.",
-			"Hosted Work execution is Workers-local; source polling is Automations.",
-			"Remote lifecycle/cancellation failures are Hosted Runner facts.",
-		),
-		nested(
-			"workers/runners/inference",
-			"workers",
-			"workers/runners",
-			"pkg/services/workers/internal/services/runners/internal/services/inference",
-			"Implement Runner for model binding/readiness/invocation through Models and output mapping.",
-			"Inference runner attempt state under Workers.",
-			"Inference attempt lifecycle coordinating Models root.",
-			"Workers Runner registry and Models Inference.",
-			"Workers owns when to infer; Models owns how inference runs.",
-			"Binding/readiness/output-mapping failures are Inference Runner facts.",
-		),
-		nested(
-			"workers/runners/script",
-			"workers",
-			"workers/runners",
-			"pkg/services/workers/internal/services/runners/internal/services/script",
-			"Implement Runner for script command policy, process execution, progress/events, timeout, and result.",
-			"Script attempt execution state under Workers.",
-			"Script runner attempt lifecycle.",
-			"Workers Runner registry.",
-			"Script attempt is Workers-local.",
-			"Timeout/process failures are Script Runner facts.",
-		),
-		nested(
-			"workers/runtime_assembly",
-			"workers",
-			"workers",
-			"pkg/services/workers/internal/services/runtime_assembly",
-			"Produce immutable executor roles from a typed Workers-owned runtime build request.",
-			"Immutable assembled executor roles/config snapshots.",
-			"Build/assembly on command; absorbs public clone/config inspection.",
-			"Factory Sessions/Runtime consumers via Workers root.",
-			"Assembly outputs are Workers-owned values, not peer constructors.",
-			"Invalid build requests are Runtime Assembly rejection facts.",
-		),
-		nested(
-			"workers/workstations",
-			"workers",
-			"workers",
-			"pkg/services/workers/internal/services/workstations",
-			"Context/environment, interpolation, prompting, optional worktree, behavior routing, dispatch, final output.",
-			"Workstation execution context for a dispatch.",
-			"Dispatch execution path lifecycle under Workers.",
-			"Runner registry and Factory Runtime dispatch consumers via Workers root.",
-			"Workstation execution shares Workers dispatch transaction.",
-			"Routing/dispatch failures are Workstation Execution facts.",
-		),
-	}
-}
 
 func committedResponsibilityClusters() []ResponsibilityCluster {
 	return []ResponsibilityCluster{
