+++ internal/ownershipinventory/rationale_factory_owners.go
@@ -1,0 +1,372 @@
+package ownershipinventory
+
+func committedAutomationsRationales() []OwnerRationaleCard {
+	return []OwnerRationaleCard{
+		topLevel(
+			"automations",
+			"pkg/services/automations",
+			"Desired/observed trigger instances, cursors/checkpoints, and Work Request generation from clocks or external observations.",
+			"Automation definitions, instance status, and cursor/checkpoint store owned by Automations.",
+			"Reconciles desired specs into long-running trigger instances with start/stop/join/restart policy.",
+			"Factory Runtime (via Work commands), Work Admission, Operator Settings, transports commanding reconcile/list/control.",
+			"Trigger instance and cursor mutations stay inside Automations; Work admission is a cross-owner command.",
+			"Missed triggers, duplicate emissions, cursor recovery, cancellation, and downstream Work rejection are typed Automations facts.",
+		),
+		nested(
+			"automations/cron",
+			"automations",
+			"automations",
+			"pkg/services/automations/internal/services/cron",
+			"Deterministic schedule ticks that command Work admission.",
+			"Schedule/jitter/expiry configuration and tick cursors under Automations store.",
+			"Produces ticks under Reconciliation supervision; no independent daemon outside Automations.",
+			"Automations Reconciliation and Work Admission via Automations root.",
+			"Shares Automations instance/cursor transaction; does not own Work content.",
+			"Invalid schedules and missed ticks surface as Automations status/recovery facts.",
+		),
+		nested(
+			"automations/filesystem_watchers",
+			"automations",
+			"automations",
+			"pkg/services/automations/internal/services/filesystem_watchers",
+			"Watch configured paths, preseed, debounce/coalesce, persist cursor, and command Work.",
+			"Watcher cursors and configured path subscriptions in Automations store.",
+			"Long-running watch instances supervised by Reconciliation.",
+			"Automations root consumers and Work Admission; moves filesystem ingest off Factory Runtime.",
+			"Cursor and subscription mutations are Automations-local; Work creation is a command.",
+			"Watcher restart and cursor recovery remain Automations failure/recovery policy.",
+		),
+		nested(
+			"automations/hosted_sources",
+			"automations",
+			"automations",
+			"pkg/services/automations/internal/services/hosted_sources",
+			"Observe hosted systems, poll/restart/checkpoint, normalize observations, and command Work.",
+			"Hosted-source cursors/checkpoints and secrets resolution handles under Automations.",
+			"Polling/restart lifecycle owned by Automations, not Workers Hosted Runner.",
+			"Automations root and Work Admission; Workers retains only hosted Work execution.",
+			"Observation/cursor state is Automations-local; Work Requests cross the Work boundary.",
+			"Poll/restart and secret-resolution failures classify as Automations source faults.",
+		),
+		nested(
+			"automations/reconciliation",
+			"automations",
+			"automations",
+			"pkg/services/automations/internal/services/reconciliation",
+			"Apply desired automation specs, construct instances, track status, and supervise lifecycle.",
+			"Desired/observed automation instance registry and status in Automations store.",
+			"Owns start/stop/join/restart orchestration for trigger subservices.",
+			"Automations root command surface and trigger subservices.",
+			"Desired-spec application and instance status mutations are one Automations aggregate.",
+			"Restart policy and reconcile conflicts remain Automations recovery behavior.",
+		),
+		nested(
+			"automations/script_pollers",
+			"automations",
+			"automations",
+			"pkg/services/automations/internal/services/script_pollers",
+			"Supervise polling commands/sources that emit canonical Work Requests.",
+			"Poller cursors/checkpoints under Automations store.",
+			"Timeout/restart supervision under Reconciliation.",
+			"Automations root and Work Admission.",
+			"Cursor and poller status are Automations-local; emitted Work crosses owners by command.",
+			"Timeout and parse failures remain Automations poller recovery facts.",
+		),
+	}
+}
+func committedFactoryDefinitionsRationales() []OwnerRationaleCard {
+	return []OwnerRationaleCard{
+		topLevel(
+			"factory_definitions",
+			"pkg/services/factory_definitions",
+			"Authored Factory aggregate catalog: validation, persistence, versioning, activation, and packaged installation.",
+			"Named Factory roots, authored layouts, .current-factory pointer, snapshots, and packaged assets.",
+			"No autonomous background daemon; activation is request-driven through Factory Sessions.",
+			"Factory Sessions, System Bootstrap, Automations (trigger specs), transports, Wire construction.",
+			"Catalog/authoring/compilation/validation/snapshot/distribution share one Definition aggregate transaction.",
+			"Malformed layouts, activation conflicts, and persistence failures remain Definition-owned typed facts.",
+		),
+		nested(
+			"factory_definitions/authoring_layout",
+			"factory_definitions",
+			"factory_definitions",
+			"pkg/services/factory_definitions/internal/services/authoring_layout",
+			"Parse/render split authored layouts and atomically create/replace/flatten/expand one Factory aggregate.",
+			"Authored layout files under Definition-owned filesystem datastore.",
+			"Request-driven authoring operations; no independent daemon.",
+			"Factory Definitions root and Definition-owned transports.",
+			"Layout writes participate in the Definition aggregate transaction.",
+			"Parse/render and layout conflicts are Definition authoring failures.",
+		),
+		nested(
+			"factory_definitions/catalog",
+			"factory_definitions",
+			"factory_definitions",
+			"pkg/services/factory_definitions/internal/services/catalog",
+			"List/delete/resolve named/project/global Factories and atomically manage the current pointer.",
+			"Named Factory catalog and .current-factory pointer.",
+			"Catalog mutations are command-driven; no background reconciler.",
+			"Factory Definitions root, System Bootstrap, Factory Sessions activation.",
+			"Catalog and current-pointer updates are atomic within Definitions.",
+			"Missing/conflicting names and pointer races are Definition catalog failures.",
+		),
+		nested(
+			"factory_definitions/compilation",
+			"factory_definitions",
+			"factory_definitions",
+			"pkg/services/factory_definitions/internal/services/compilation",
+			"Convert authored/canonical source into one normalized LoadedFactorySource without executing it.",
+			"Compiled effective definition artifacts derived from authored store.",
+			"Compile on command; Runtime executes, Definitions does not.",
+			"Factory Definitions root and Runtime semantic-validation callers via root.",
+			"Compilation outputs are Definition-owned; Runtime does not mutate authored store.",
+			"Compile/reference-resolution failures stay on the Definition boundary.",
+		),
+		nested(
+			"factory_definitions/distribution",
+			"factory_definitions",
+			"factory_definitions",
+			"pkg/services/factory_definitions/internal/services/distribution",
+			"Built-in package catalog, packaged installation, and scaffold creation producing Factory aggregates.",
+			"Packaged Factory assets and installation outputs in Definition store.",
+			"Install/scaffold commands; no marketplace lifecycle in this program.",
+			"System Bootstrap, CLI/HTTP Definition adapters, Factory Definitions root.",
+			"Installation writes the same Definition aggregate transactionally.",
+			"Install/scaffold failures roll back or report Definition distribution faults.",
+		),
+		nested(
+			"factory_definitions/invocation_policy",
+			"factory_definitions",
+			"factory_definitions",
+			"pkg/services/factory_definitions/internal/services/invocation_policy",
+			"Definition-owned interpretation of authored Factory behavior at invocation time: decision envelopes, interpolation, output shaping, work-type selection, quorum lineage, work propagation, workstation execution limits, and TTS observability identity.",
+			"Stateless invocation-time policy derivations over authored worker, workstation, and packaged configuration.",
+			"Policy interpretation runs on command during worker and workstation invocation; no independent daemon.",
+			"Factory Definitions root, Factory Sessions invocation, and Workers dispatch consumers via Definitions root.",
+			"Invocation policy does not mutate peer stores; interpretation outputs are Definition-owned values consumed at invocation.",
+			"Malformed authored policy and unsupported invocation contracts surface as Definition-owned rejection facts.",
+		),
+		nested(
+			"factory_definitions/snapshots_portability",
+			"factory_definitions",
+			"factory_definitions",
+			"pkg/services/factory_definitions/internal/services/snapshots_portability",
+			"Capture detached snapshots and prepare/materialize bundled assets for export/import/replay use.",
+			"Editable/version snapshots and bundled portable assets.",
+			"Snapshot/export/import are request-driven operations.",
+			"Recordings replay consumers via Definition compile/import root operations; Definition transports.",
+			"Snapshot materialization stays inside Definitions; Recordings owns replay plans.",
+			"Portability validation and materialization failures are Definition-owned.",
+		),
+		nested(
+			"factory_definitions/validation",
+			"factory_definitions",
+			"factory_definitions",
+			"pkg/services/factory_definitions/internal/services/validation",
+			"Structural, topology, pre-persist, tool, and workstation compatibility validation policy.",
+			"Validation policy and diagnostics over Definition aggregate; may query Runtime root for orchestration semantics.",
+			"Validation runs on command before persist/activation.",
+			"Factory Definitions root and authoring/activation callers.",
+			"Validation does not mutate peer stores; persist remains Definition transaction.",
+			"Validation diagnostics are Definition rejection facts, not Runtime failures.",
+		),
+	}
+}
+func committedFactoryRuntimeRationales() []OwnerRationaleCard {
+	return []OwnerRationaleCard{
+		topLevel(
+			"factory_runtime",
+			"pkg/services/factory_runtime",
+			"Per-session orchestration, execution state, instance mechanics, dispatch planning, and checkpoints.",
+			"Runtime checkpoints, submission/result/dispatch buffers, and live instance state.",
+			"Owns run handle pause/resume/stop/replacement/recovery for one leased execution instance.",
+			"Factory Sessions (lease/start), Workers (dispatch), Work, Recordings (event candidates), Visualization (sanitized observations).",
+			"Runtime instance and checkpoint mutations are Runtime-local; peer Work/Workers/Recordings interactions are commands/events.",
+			"Crash recovery uses Runtime checkpoints plus Recordings history; orchestration failures normalize at Runtime root.",
+		),
+		nested(
+			"factory_runtime/checkpoint_recovery",
+			"factory_runtime",
+			"factory_runtime",
+			"pkg/services/factory_runtime/internal/services/checkpoint_recovery",
+			"Capture/load versioned execution checkpoints and restore orchestration state idempotently.",
+			"Runtime-owned checkpoint store coordinated with Recordings history ownership.",
+			"Recovery commands after durability decision; not an independent deployable yet.",
+			"Factory Runtime root and Instance Host.",
+			"Checkpoint write/load is Runtime-local; canonical history remains Recordings.",
+			"Incompatible checkpoints and restore conflicts are Runtime recovery failures.",
+		),
+		nested(
+			"factory_runtime/dispatch_planning",
+			"factory_runtime",
+			"factory_runtime",
+			"pkg/services/factory_runtime/internal/services/dispatch_planning",
+			"Convert enabled transitions into stable dispatch intents and accept correlated results.",
+			"In-flight dispatch intent/outbox buffers owned by Runtime.",
+			"Publishes intents after Workers command/result contract lock.",
+			"Workers execution and Factory Runtime Instance Host.",
+			"Intent lifecycle is Runtime-local; Workers owns execution attempts.",
+			"Duplicate/correlated-result conflicts recover inside Runtime outbox policy.",
+		),
+		nested(
+			"factory_runtime/instance_host",
+			"factory_runtime",
+			"factory_runtime",
+			"pkg/services/factory_runtime/internal/services/instance_host",
+			"Construct inert instances from lease/start data and manage run handle lifecycle.",
+			"Live instance handles and runtime-owned artifact closure metadata.",
+			"Start/stop/pause/resume/replace/recover for one Runtime instance.",
+			"Factory Sessions placement/lease commands and Orchestration.",
+			"Instance handle mutations stay in Runtime; Session desired lifecycle stays in Sessions.",
+			"Replacement and recovery failures remain Runtime instance faults.",
+		),
+		nested(
+			"factory_runtime/orchestration",
+			"factory_runtime",
+			"factory_runtime",
+			"pkg/services/factory_runtime/internal/services/orchestration",
+			"Select orchestration kind, compile/validate/preview/execute/resume strategy semantics, encode checkpoints.",
+			"Compiled orchestration strategy state for the active instance.",
+			"Execute/resume under Instance Host; Petri/JavaScript are internal variants.",
+			"Factory Runtime root, Definition semantic validation callers via Runtime root.",
+			"Strategy execution shares Runtime instance transaction; authored Factory config remains Definitions.",
+			"Strategy validation and execution faults normalize through Runtime orchestration.",
+		),
+	}
+}
+func committedFactorySessionsRationales() []OwnerRationaleCard {
+	return []OwnerRationaleCard{
+		topLevel(
+			"factory_sessions",
+			"pkg/services/factory_sessions",
+			"Session identity, desired lifecycle, placement, recovery, discovery, invocation coordination, and session response streams.",
+			"Live session registry, session control store, and session-owned persistence coordination (canonical history delegated to Recordings).",
+			"Owns desired lifecycle reconcile, live selection/activation, and durable start/resume/control.",
+			"Transports, Factory Runtime, Work, Workers, Provider Sessions, Models, Visualization, Recordings.",
+			"Session aggregate mutations are Sessions-local; peer services are commanded through roots.",
+			"Placement/activation conflicts, durable control failures, and timeout/cancellation are Sessions recovery facts.",
+		),
+		nested(
+			"factory_sessions/durable_execution",
+			"factory_sessions",
+			"factory_sessions",
+			"pkg/services/factory_sessions/internal/services/durable_execution",
+			"Durable start/resume/control/inspection and session-owned persistence coordination.",
+			"Session durable execution identity/persistence handles; delegates canonical history/artifacts to Recordings.",
+			"Durable control lifecycle for a session.",
+			"Factory Sessions root, Live Runtime, Recordings root.",
+			"Session persistence coordination is Sessions-local; ledger append is Recordings command.",
+			"Durable control and resume conflicts recover in Durable Execution.",
+		),
+		nested(
+			"factory_sessions/identity",
+			"factory_sessions",
+			"factory_sessions",
+			"pkg/services/factory_sessions/internal/services/identity",
+			"Normalize/discover targets, derive logical keys, resolve folders, registry-aware identity lookup.",
+			"Logical session keys and folder/identity resolution metadata.",
+			"Identity resolution is request-driven under Sessions.",
+			"Factory Sessions root and nested Live Runtime/Durable Execution.",
+			"Identity derivation shares Sessions aggregate authority.",
+			"Ambiguous/missing identity targets are Sessions identity failures.",
+		),
+		nested(
+			"factory_sessions/invocation",
+			"factory_sessions",
+			"factory_sessions",
+			"pkg/services/factory_sessions/internal/services/invocation",
+			"Resolve invocation input, command Work, observe completion, enforce timeout/cancellation, record telemetry.",
+			"Per-session invocation observation state.",
+			"Timeout/cancellation lifecycle for one session invocation.",
+			"Work root, Factory Sessions root, Recordings telemetry consumers.",
+			"Invocation coordination is Sessions-local; Work admission/state are Work commands.",
+			"Timeouts and cancellations are Sessions invocation recovery facts.",
+		),
+		nested(
+			"factory_sessions/live_runtime",
+			"factory_sessions",
+			"factory_sessions",
+			"pkg/services/factory_sessions/internal/services/live_runtime",
+			"Open/list/get/snapshot/pause/resume/close; sole owner of live registry, selection, activation lock, cleanup.",
+			"Live session registry and mutable live-runtime aggregate.",
+			"Live activation lock and cleanup lifecycle.",
+			"Factory Sessions root, Runtime Opening, Durable Execution, Response Stream.",
+			"Live registry mutations are exclusive to Live Runtime inside Sessions.",
+			"Activation races and cleanup failures recover under Live Runtime.",
+		),
+		nested(
+			"factory_sessions/response_stream",
+			"factory_sessions",
+			"factory_sessions",
+			"pkg/services/factory_sessions/internal/services/response_stream",
+			"Allocate session event stores, publish/complete, filter subscriptions, enforce retention/backpressure, track cursors.",
+			"Ephemeral session response-event stores and cursors.",
+			"Publish/complete and subscription lifecycle for a session.",
+			"Factory Sessions root, transports/Visualization presentation consumers.",
+			"Response-store mutations are Sessions-local; canonical Factory events belong to Recordings.",
+			"Backpressure/drop and retention faults are Response Stream recovery policy.",
+		),
+		nested(
+			"factory_sessions/runtime_opening",
+			"factory_sessions",
+			"factory_sessions",
+			"pkg/services/factory_sessions/internal/services/runtime_opening",
+			"Coordinate peer root services and return a lifecycle plan without importing peer implementations.",
+			"No independent datastore; produces lifecycle plan over already-constructed roles.",
+			"Opening/binding lifecycle after root binding contract stabilizes.",
+			"Factory Sessions root and peer service roots (Definitions, Runtime, Workers, Models, Recordings).",
+			"Does not own peer transactions; only plans Sessions-owned activation steps.",
+			"Binding/plan failures surface as Sessions opening faults without peer downcasts.",
+		),
+	}
+}
+func committedFactoryVisualizationRationales() []OwnerRationaleCard {
+	return []OwnerRationaleCard{
+		topLevel(
+			"factory_visualization",
+			"pkg/services/factory_visualization",
+			"Request-activated visualization lifecycle, live projection, response/event presentation, queues, backpressure, final-write coordination.",
+			"Visualization cursors, output queues, and presentation state for an activated request.",
+			"Inert by default; Start/Stop/Wait/cleanup only when request parameters activate it.",
+			"Transports (parameters/codecs/sinks), Recordings query/subscription, sanitized Runtime observations.",
+			"Activation and presentation mutations are Visualization-local; Recordings/Runtime remain peer roots.",
+			"Backpressure drops, activation failures, and drain errors are Visualization recovery facts.",
+		),
+		nested(
+			"factory_visualization/activation_lifecycle",
+			"factory_visualization",
+			"factory_visualization",
+			"pkg/services/factory_visualization/internal/services/activation_lifecycle",
+			"Interpret approved request options, bind session/runtime, own Start/Stop/Wait/cleanup.",
+			"Activation binding state for the request.",
+			"Request-driven visualization lifecycle.",
+			"Factory Visualization root and transport adapters.",
+			"Activation state is Visualization-local.",
+			"Failed binds and unclean shutdowns are activation recovery faults.",
+		),
+		nested(
+			"factory_visualization/live_view_projection",
+			"factory_visualization",
+			"factory_visualization",
+			"pkg/services/factory_visualization/internal/services/live_view_projection",
+			"Retain cursor/events, subscribe retained-then-live, consume Runtime observations and Recordings queries, emit View.",
+			"Live view cursor/event retention owned by Visualization.",
+			"Subscription lifecycle under activation.",
+			"Recordings projection/query root and Runtime observation contract.",
+			"View projection state is Visualization-local; canonical projections stay in Recordings.",
+			"Reconnect/gap handling follows Visualization cursor policy.",
+		),
+		nested(
+			"factory_visualization/response_event_presentation",
+			"factory_visualization",
+			"factory_visualization",
+			"pkg/services/factory_visualization/internal/services/response_event_presentation",
+			"Best-effort/lossless queues, bounded drop/backpressure, exclusive final writes, drain.",
+			"Output queues and presentation buffers.",
+			"Queue drain lifecycle tied to activation.",
+			"Transport sinks/codecs and Visualization root.",
+			"Queue mutations are Visualization-local.",
+			"Drop/backpressure and final-write conflicts are presentation recovery facts.",
+		),
+	}
+}
